requirements.txt
opencv-python==4.12.0.88
numpy==2.1.2
insightface==0.7.3
paho-mqtt==2.1.0
